primary-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
exceptiongroup==1.3.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.2.0
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
youtube-search-python==1.6.6
